pyproject.toml
[project]
name = "nablafx"
version = "1.0.0"
description = "Differentiable black-box and gray-box audio effects modeling framework"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {file = "LICENSE"}

authors = [
    {name = "Marco Comunità", email = "m.comunita@qmul.ac.uk"},
]

maintainers = [
    {name = "Marco Comunità", email = "m.comunita@qmul.ac.uk"},
]

keywords = [
    "audio",
    "audio-effects",
    "deep-learning",
    "differentiable-dsp",
    "pytorch",
    "audio-processing",
    "neural-audio",
    "black-box-modeling",
    "gray-box-modeling",
    "machine-learning",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "auraloss>=0.4.0",
    "einops>=0.8.0",
    "frechet-audio-distance",
    "lightning>=2.4",
    "lightning-utilities>=0.11.3",
    "matplotlib>=3.9",
    "natsort",
    "numpy>=1.26,<3",
    "scipy>=1.13",
    "torch>=2.8",
    "torchaudio>=2.8",
    "torchmetrics>=1.4",
    "torchvision>=0.23",
    "typing-extensions>=4.12",
    "wandb>=0.17",
    "rational-activations==0.2.0",
    "torchcodec>=0.11.1",
    # NPP (image primitives) isn't bundled by the torch cu130 wheel but is
    # linked at load time by torchcodec. Only Linux wheels exist.
    "nvidia-npp>=13.1; sys_platform == 'linux'",
    "hydra-core>=1.3",
]


[project.scripts]
nablafx = "nablafx.__main__:main"
nablafx-export = "nablafx.export.cli:main"
nablafx-plugin = "nablafx.plugin.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project.optional-dependencies]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.2",
]

# ONNX stack is only needed by the plugin-export pipeline, not by training.
# Install with: uv sync --extra export
export = [
    "onnx>=1.17",
    "onnxruntime>=1.20",
    "pyyaml>=6.0",
]

# PEP 735 groups — installed by default with `uv sync`, not shipped in the wheel.
[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/mcomunita/nablafx"
Documentation = "https://github.com/mcomunita/nablafx#readme"
Repository = "https://github.com/mcomunita/nablafx.git"
"Bug Tracker" = "https://github.com/mcomunita/nablafx/issues"
"Installation Guide" = "https://github.com/mcomunita/nablafx/blob/master/INSTALL.md"
"Paper" = "https://arxiv.org/abs/2502.11668"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["nablafx*"]
exclude = ["tests*", "benchmarks*", "scripts*", "docs*"]

[tool.setuptools.package-data]
nablafx = ["py.typed", "rationals_config.json"]

# --- uv-specific configuration -------------------------------------------------
#
# rational-activations 0.2.0 declares `torch==1.7.1` in its metadata, but the
# wheel is pure-Python and works with any modern torch. We override that pin
# globally so uv's resolver doesn't try to pull torch 1.7.1. We're stuck on
# 0.2.0 because 0.2.1 dropped the universal `py3-none-any` wheel in favor of
# cp36/37/38/39-only C-extension wheels with no sdist, which don't cover
# Python >= 3.10.
#
# torch/torchaudio/torchvision are pulled from the PyTorch CUDA 13 wheel index
# for native Blackwell (sm_121) / CUDA 13 support (DGX Spark).
#
[tool.uv]
override-dependencies = [
    "torch>=2.8",
]

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchaudio = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
